Multithreading functional

+++ miniRT/srcs/lib_math/vec3_norm.c
@@ -1,8 +1,8 @@
 #include "miniRT.h"
 
-float	get_norm2_vec3(t_vec3 vector)
+double	get_norm2_vec3(t_vec3 vector)
 {
-	float	norm;
+	double	norm;
 
 	norm = vector.coord[X] * vector.coord[X]
 		+ vector.coord[Y] * vector.coord[Y]
@@ -10,9 +10,9 @@ float	get_norm2_vec3(t_vec3 vector)
 	return (norm);
 }
 
-float	get_norm_vec3(t_vec3 vector)
+double	get_norm_vec3(t_vec3 vector)
 {
-	float	norm;
+	double	norm;
 
 	norm = sqrt(get_norm2_vec3(vector));
 	return (norm);
